Support nodepools under machinepool commands
