+++ cmd/dlt/machinepool/cmd.go
@@ -18,20 +18,12 @@ package machinepool
 
 import (
 	"fmt"
-	"os"
-	"regexp"
 
-	cmv1 "github.com/openshift-online/ocm-sdk-go/clustersmgmt/v1"
 	"github.com/spf13/cobra"
 
-	"github.com/openshift/rosa/pkg/interactive/confirm"
 	"github.com/openshift/rosa/pkg/ocm"
 	"github.com/openshift/rosa/pkg/rosa"
 )
-
-// Regular expression to used to make sure that the identifier given by the
-// user is safe and that it there is no risk of SQL injection:
-var machinePoolKeyRE = regexp.MustCompile(`^[a-z]([-a-z0-9]*[a-z0-9])?$`)
 
 var Cmd = &cobra.Command{
 	Use:     "machinepool ID",
@@ -60,46 +52,12 @@ func run(_ *cobra.Command, argv []string) {
 	defer r.Cleanup()
 
 	machinePoolID := argv[0]
-	if machinePoolID != "Default" && !machinePoolKeyRE.MatchString(machinePoolID) {
-		r.Reporter.Errorf("Expected a valid identifier for the machine pool")
-		os.Exit(1)
-	}
-
 	clusterKey := r.GetClusterKey()
-
-	if machinePoolID == "Default" {
-		r.Reporter.Errorf("Machine pool '%s' cannot be deleted from cluster '%s'", machinePoolID, clusterKey)
-		os.Exit(1)
-	}
-
 	cluster := r.FetchCluster()
-	// Try to find the machine pool:
-	r.Reporter.Debugf("Loading machine pools for cluster '%s'", clusterKey)
-	machinePools, err := r.OCMClient.GetMachinePools(cluster.ID())
-	if err != nil {
-		r.Reporter.Errorf("Failed to get machine pools for cluster '%s': %v", clusterKey, err)
-		os.Exit(1)
-	}
 
-	var machinePool *cmv1.MachinePool
-	for _, item := range machinePools {
-		if item.ID() == machinePoolID {
-			machinePool = item
-		}
-	}
-	if machinePool == nil {
-		r.Reporter.Errorf("Failed to get machine pool '%s' for cluster '%s'", machinePoolID, clusterKey)
-		os.Exit(1)
-	}
-
-	if confirm.Confirm("delete machine pool '%s' on cluster '%s'", machinePoolID, clusterKey) {
-		r.Reporter.Debugf("Deleting machine pool '%s' on cluster '%s'", machinePool.ID(), clusterKey)
-		err = r.OCMClient.DeleteMachinePool(cluster.ID(), machinePool.ID())
-		if err != nil {
-			r.Reporter.Errorf("Failed to delete machine pool '%s' on cluster '%s': %s",
-				machinePool.ID(), clusterKey, err)
-			os.Exit(1)
-		}
-		r.Reporter.Infof("Successfully deleted machine pool '%s' from cluster '%s'", machinePoolID, clusterKey)
+	if cluster.Hypershift().Enabled() {
+		deleteNodePool(r, machinePoolID, clusterKey, cluster)
+	} else {
+		deleteMachinePool(r, machinePoolID, clusterKey, cluster)
 	}
 }
