+++ cmd/edit/machinepool/nodepool.go
@@ -1,0 +1,152 @@
+package machinepool
+
+import (
+	"os"
+
+	cmv1 "github.com/openshift-online/ocm-sdk-go/clustersmgmt/v1"
+	"github.com/openshift/rosa/pkg/interactive"
+	rprtr "github.com/openshift/rosa/pkg/reporter"
+	"github.com/openshift/rosa/pkg/rosa"
+	"github.com/spf13/cobra"
+)
+
+func editNodePool(cmd *cobra.Command, nodePoolID string, clusterKey string, cluster *cmv1.Cluster, r *rosa.Runtime) {
+	var err error
+
+	// Try to find the node pool
+	r.Reporter.Debugf("Loading machine pool for hosted cluster '%s'", clusterKey)
+	nodePool, err := r.OCMClient.GetNodePool(cluster.ID(), nodePoolID)
+	if err != nil {
+		r.Reporter.Errorf("Failed to get machine pools for hosted cluster '%s': %v", clusterKey, err)
+		os.Exit(1)
+	}
+
+	autoscaling, replicas, minReplicas, maxReplicas := getNodePoolReplicas(cmd, r.Reporter, nodePoolID,
+		nodePool.Replicas(), nodePool.Autoscaling())
+
+	if !autoscaling && replicas < 0 ||
+		(autoscaling && cmd.Flags().Changed("min-replicas") && minReplicas < 0) {
+		r.Reporter.Errorf("The number of machine pool replicas needs to be a non-negative integer")
+		os.Exit(1)
+	}
+
+	npBuilder := cmv1.NewNodePool().
+		ID(nodePool.ID())
+
+	if autoscaling {
+		asBuilder := cmv1.NewNodePoolAutoscaling()
+
+		if minReplicas > 0 {
+			asBuilder = asBuilder.MinReplica(minReplicas)
+		}
+		if maxReplicas > 0 {
+			asBuilder = asBuilder.MaxReplica(maxReplicas)
+		}
+
+		npBuilder = npBuilder.Autoscaling(asBuilder)
+	} else {
+		npBuilder = npBuilder.Replicas(replicas)
+	}
+
+	nodePool, err = npBuilder.Build()
+	if err != nil {
+		r.Reporter.Errorf("Failed to create machine pool for hosted cluster '%s': %v", clusterKey, err)
+		os.Exit(1)
+	}
+
+	r.Reporter.Debugf("Updating machine pool '%s' on hosted cluster '%s'", nodePool.ID(), clusterKey)
+	_, err = r.OCMClient.UpdateNodePool(cluster.ID(), nodePool)
+	if err != nil {
+		r.Reporter.Errorf("Failed to update machine pool '%s' on hosted cluster '%s': %s",
+			nodePool.ID(), clusterKey, err)
+		os.Exit(1)
+	}
+	r.Reporter.Infof("Updated machine pool '%s' on hosted cluster '%s'", nodePool.ID(), clusterKey)
+}
+
+func getNodePoolReplicas(cmd *cobra.Command,
+	reporter *rprtr.Object,
+	nodePoolID string,
+	existingReplicas int,
+	existingAutoscaling *cmv1.NodePoolAutoscaling) (autoscaling bool, replicas, minReplicas, maxReplicas int) {
+	var err error
+
+	isMinReplicasSet := cmd.Flags().Changed("min-replicas")
+	isMaxReplicasSet := cmd.Flags().Changed("max-replicas")
+	isReplicasSet := cmd.Flags().Changed("replicas")
+	isAutoscalingSet := cmd.Flags().Changed("enable-autoscaling")
+
+	replicas = args.replicas
+	minReplicas = args.minReplicas
+	maxReplicas = args.maxReplicas
+	autoscaling = args.autoscalingEnabled
+	replicasRequired := existingAutoscaling == nil
+
+	// if the user set min/max replicas and hasn't enabled autoscaling, or existing is disabled
+	if (isMinReplicasSet || isMaxReplicasSet) && !autoscaling && existingAutoscaling == nil {
+		reporter.Errorf("Autoscaling is not enabled on machine pool '%s'. can't set min or max replicas",
+			nodePoolID)
+		os.Exit(1)
+	}
+
+	if !isAutoscalingSet {
+		autoscaling = existingAutoscaling != nil
+		if interactive.Enabled() {
+			autoscaling, err = interactive.GetBool(interactive.Input{
+				Question: "Enable autoscaling",
+				Help:     cmd.Flags().Lookup("enable-autoscaling").Usage,
+				Default:  autoscaling,
+				Required: false,
+			})
+			if err != nil {
+				reporter.Errorf("Expected a valid value for enable-autoscaling: %s", err)
+				os.Exit(1)
+			}
+		}
+	}
+
+	if autoscaling {
+		// Prompt for min replicas if neither min or max is set or interactive mode
+		if !isMinReplicasSet && (interactive.Enabled() || !isMaxReplicasSet) {
+			minReplicas, err = interactive.GetInt(interactive.Input{
+				Question: "Min replicas",
+				Help:     cmd.Flags().Lookup("min-replicas").Usage,
+				Default:  existingAutoscaling.MinReplica(),
+				Required: replicasRequired,
+			})
+			if err != nil {
+				reporter.Errorf("Expected a valid number of min replicas: %s", err)
+				os.Exit(1)
+			}
+		}
+
+		// Prompt for max replicas if neither min or max is set or interactive mode
+		if !isMaxReplicasSet && (interactive.Enabled() || !isMinReplicasSet) {
+			maxReplicas, err = interactive.GetInt(interactive.Input{
+				Question: "Max replicas",
+				Help:     cmd.Flags().Lookup("max-replicas").Usage,
+				Default:  existingAutoscaling.MaxReplica(),
+				Required: replicasRequired,
+			})
+			if err != nil {
+				reporter.Errorf("Expected a valid number of max replicas: %s", err)
+				os.Exit(1)
+			}
+		}
+	} else if interactive.Enabled() || !isReplicasSet {
+		if !isReplicasSet {
+			replicas = existingReplicas
+		}
+		replicas, err = interactive.GetInt(interactive.Input{
+			Question: "Replicas",
+			Help:     cmd.Flags().Lookup("replicas").Usage,
+			Default:  replicas,
+			Required: true,
+		})
+		if err != nil {
+			reporter.Errorf("Expected a valid number of replicas: %s", err)
+			os.Exit(1)
+		}
+	}
+	return
+}
